shallow copy of site's context to avoid modifying
when initializaing the page's context with the site's we need to make a
shallow copy to avoid updating the site's (global) context cache with the
page specific information. ran in to this when adding
    title: Page Specific Title
to pages and noticed that subsequent page's would get the previous page's
title if it was omitted.